requirements.txt
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
cryptography==2.9
daphne==2.4.1
Django==3.0.4
django-jsonfield==1.4.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2019.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.0.2
